requirements.txt
lychee
kivy
Flask
python-dotenv
